refactoring

+++ src/main/java/com/c446/ironbound_core/entity/spells/moonlightBeam/MoonlightRay.java
@@ -65,7 +65,7 @@ public class MoonlightRay extends Projectile implements AntiMagicSusceptible {
     }
 
     public float getRadius() {
-        return (Float) this.getEntityData().get(DATA_RADIUS);
+        return this.getEntityData().get(DATA_RADIUS);
     }
 
     @Override
